third_party/blink/web_tests/external/wpt/css/cssom/font-variant-shorthand-serialization-expected.txt

This is a testharness.js-based test.
[PASS] font-variant: normal serialization
[PASS] font-variant: none serialization
[PASS] font-variant-ligatures: none serialization with non-default value for another longhand
[PASS] font-variant: normal with non-default longhands
[PASS] CSS-wide keyword in one longhand
[PASS] CSS-wide keyword in shorthand
[FAIL] font: menu serialization
  assert_equals: font-variant was set expected "" but got "normal"
Harness: the test ran to completion.
